[nrfconnect] Introduce the WiFi connection recovery feature
The WiFi connection recovery feature allows device re-scanning
and re-connecting to the known WiFi network after the device's reboot
and when the known SSID has not been found during the last scan.
The connection recovery interval is doubled with
every occurrence to the defined maximum value and then its value
depends on the maximum value +- the defined random jitter.

After restoring the connection, the Connection Recovery
Interval is restored to the defined minimum value after elapsing
of the defined delay to avoid frequent reconnections to a
poor link quality network.